+++ mimic-bukkit/src/main/kotlin/command/MainCommand.kt
@@ -1,49 +1,48 @@
 package ru.endlesscode.mimic.command
 
-import co.aikar.commands.CommandHelp
-import co.aikar.commands.MimicCommand
-import co.aikar.commands.annotation.*
+import dev.jorel.commandapi.executors.CommandExecutor
+import dev.jorel.commandapi.kotlindsl.commandAPICommand
 import net.kyori.adventure.platform.bukkit.BukkitAudiences
 import net.kyori.adventure.text.format.NamedTextColor
-import org.bukkit.command.CommandSender
-import org.bukkit.plugin.Plugin
+import ru.endlesscode.mimic.Mimic
+import ru.endlesscode.mimic.config.MimicConfig
 import ru.endlesscode.mimic.internal.append
 import ru.endlesscode.mimic.internal.appendClickable
 import ru.endlesscode.mimic.internal.appendLine
 import ru.endlesscode.mimic.internal.buildTextComponent
 
-@CommandAlias("%command")
-@CommandPermission("%perm")
-internal class MainCommand(
-    private val plugin: Plugin,
-    private val audiences: BukkitAudiences,
-) : MimicCommand() {
+/** Registers command '/mimic' and all subcommands. */
+internal fun registerCommand(
+    mimic: Mimic,
+    config: MimicConfig,
+    pluginFullName: String,
+    audiences: BukkitAudiences,
+) = commandAPICommand("mimic") {
+    withPermission("mimic.admin")
+    withShortDescription("Show info about Mimic")
+    executes(infoExecutor(audiences, pluginFullName))
 
-    @HelpCommand
-    @Description("Show help")
-    fun doHelp(help: CommandHelp) {
-        help.showHelp()
-    }
-
-    @Subcommand("info")
-    @Description("Show info about Mimic and loaded services")
-    fun info(sender: CommandSender) {
-        val message = buildTextComponent {
-            appendLine(plugin.description.fullName, NamedTextColor.GREEN)
-            color(NamedTextColor.GRAY)
-            append("Use ")
-            append(createClickableCommand())
-            append(" to see or change configs")
-        }
-        audiences.sender(sender).sendMessage(message)
-    }
-
-    private fun createClickableCommand() = buildTextComponent {
-        color(NamedTextColor.YELLOW)
-        appendClickable(CONFIG_COMMAND, "Click to execute", CONFIG_COMMAND)
-    }
-
-    private companion object {
-        const val CONFIG_COMMAND = "/mimic config"
-    }
+    configSubcommand(mimic, config, audiences)
+    levelSystemSubcommand(mimic)
+    classSystemSubcommand(mimic)
+    inventorySubcommand(mimic)
+    itemsSubcommand(mimic.getItemsRegistry())
 }
+
+private fun infoExecutor(audiences: BukkitAudiences, pluginFullName: String) = CommandExecutor { sender, _ ->
+    val message = buildTextComponent {
+        appendLine(pluginFullName, NamedTextColor.GREEN)
+        color(NamedTextColor.GRAY)
+        append("Use ")
+        append(createClickableCommand())
+        append(" to see or change configs")
+    }
+    audiences.sender(sender).sendMessage(message)
+}
+
+private fun createClickableCommand() = buildTextComponent {
+    color(NamedTextColor.YELLOW)
+    appendClickable(CONFIG_COMMAND, "Click to execute", CONFIG_COMMAND)
+}
+
+private const val CONFIG_COMMAND = "/mimic config"